Modified methods that encode generated palette images
Exchanged the palette.show() method in favor of palette.save() and palette.close(). This allows me to save the image as a .JPEG, which is compatible with Imgur, unlike the default .BMP format images are encoded in. These methods also temporarily save the image to local storage so that they may be uploaded to Imgur.